backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.14.0
qdrant-client==1.16.2
sqlalchemy==2.0.27
asyncpg==0.29.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
python-frontmatter==1.0.1
tiktoken==0.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
